pyproject.toml
[project]
name = "casino-virtual"
version = "0.1.0"
description = ""
authors = [
    {name = "sga-encoder",email = "sebastian.garzon54795@ucaldas.edu.co"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firebase-admin (>=6.8.0,<7.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "google-cloud-firestore (>=2.20.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "src"}]
